Add mobile test case
