+++ composer.json
@@ -2,7 +2,7 @@
   "name": "sauce/sausage",
   "type": "sauce-sausage",
   "description": "PHP version of the Sauce Labs API",
-  "keywords": ["sauce", "saucelabs", "api", "testing", "phpunit", "selenium"],
+  "keywords": ["sauce", "saucelabs", "api", "testing", "phpunit", "selenium", "appium"],
   "homepage": "http://github.com/jlipps/sausage",
   "license": "Apache-2.0",
   "minimum-stability": "dev",
@@ -13,11 +13,18 @@
     "homepage": "http://www.saucelabs.com",
     "role": "Lead"
   }],
+  "repositories": [
+      {
+          "type": "vcs",
+          "url": "https://github.com/appium/php-client"
+      }
+  ],
   "require": {
     "php": ">=5.3.0",
     "sauce/sausage-installer": ">=0.1.0",
     "phpunit/phpunit-selenium": ">=1.3.3",
-    "brianium/paratest": ">=0.6.1"
+    "brianium/paratest": ">=0.6.1",
+    "appium/php-client": "dev-master"
   },
   "suggest": {
     "sauce/connect": ">=3.0"
