+++ src/Sauce/Sausage/MobileTestCase.php
@@ -1,0 +1,7 @@
+<?php
+namespace Sauce\Sausage;
+
+abstract class MobileTestCase extends \PHPUnit_Extensions_AppiumTestCase
+{
+    use TestCase;
+}
